+++ src/module/config.mjs
@@ -13,7 +13,7 @@ import { preLocalize } from "./helpers/localization.mjs";
  * @remarks "none" is reserved for cases where we want an explicit non-option *and* default fallbacks
  * @type {Record<string, {label: string; hint: string; rollKey: string}>}
  */
-export const characteristics = {
+export const characteristics = Object.seal({
   might: {
     label: "DRAW_STEEL.Actor.characteristics.might.full",
     hint: "DRAW_STEEL.Actor.characteristics.might.abbreviation",
@@ -39,7 +39,7 @@ export const characteristics = {
     hint: "DRAW_STEEL.Actor.characteristics.presence.abbreviation",
     rollKey: "P",
   },
-};
+});
 preLocalize("characteristics", { keys: ["label", "hint"] });
 
 /* -------------------------------------------------- */